refactor (server/cli): update the cliParser have more specific error case

+++ zappy/src/zappy/server/CliParser.hpp
@@ -32,7 +32,7 @@ class CliParser {
         std::uint32_t mapHeight = 0;
         std::vector<std::string_view> teamsName;
         std::uint32_t nbInitialClient = 0;
-        std::uint8_t frequencies = 0;
+        std::uint16_t frequencies = 0;
     };
 
     const CliParameters& parameters();
